refactor and add edit cart item e2e test
