+++ e2e/tests/desktop/registered-shopper.spec.js
@@ -7,6 +7,7 @@
 
 const { test, expect } = require("@playwright/test");
 const config = require("../../config");
+const { addProductToCart } = require("../../scripts/pageHelpers");
 const {
   generateUserCredentials,
   getCreditCardExpiry,
@@ -46,54 +47,8 @@ test("Registered shopper can checkout items", async ({ page }) => {
   await expect(page.getByText(REGISTERED_USER_CREDENTIALS.email)).toBeVisible();
 
   // Shop for items as registered user
-  await page.goto(config.RETAIL_APP_HOME);
+  await addProductToCart(page)
 
-  await page.getByRole("link", { name: "Womens" }).hover();
-  const topsNav = await page.getByRole("link", { name: "Tops", exact: true });
-  await expect(topsNav).toBeVisible();
-
-  await topsNav.click();
-
-  // PLP
-  const productTile = page.getByRole("link", {
-    name: /Cotton Turtleneck Sweater/i,
-  });
-  // selecting swatch
-  const productTileImg = productTile.locator("img");
-  await productTileImg.waitFor({state: 'visible'})
-  const initialSrc = await productTileImg.getAttribute("src");
-  await expect(productTile.getByText(/From \$39\.99/i)).toBeVisible();
-
-  await productTile.getByLabel(/Black/, { exact: true }).hover();
-  // Make sure the image src has changed
-  await expect(async () => {
-    const newSrc = await productTileImg.getAttribute("src")
-    expect(newSrc).not.toBe(initialSrc)
-  }).toPass()
-  await expect(productTile.getByText(/From \$39\.99/i)).toBeVisible();
-  await productTile.click();
-
-  // PDP
-  await expect(
-    page.getByRole("heading", { name: /Cotton Turtleneck Sweater/i })
-  ).toBeVisible();
-  await page.getByRole("radio", { name: "L", exact: true }).click();
-
-  await page.locator("button[data-testid='quantity-increment']").click();
-
-  // // Selected Size and Color texts are broken into multiple elements on the page.
-  // // So we need to look at the page URL to verify selected variants
-  const updatedPageURL = await page.url();
-  const params = updatedPageURL.split("?")[1];
-  expect(params).toMatch(/size=9LG/i);
-  expect(params).toMatch(/color=JJ169XX/i);
-  await page.getByRole("button", { name: /Add to Cart/i }).click();
-
-  const addedToCartModal = page.getByText(/2 items added to cart/i);
-
-  await addedToCartModal.waitFor();
-
-  await page.getByLabel("Close").click();
   // cart
   await page.getByLabel(/My cart/i).click();
 
@@ -194,3 +149,5 @@ test("Registered shopper can checkout items", async ({ page }) => {
     page.getByRole("link", { name: /Cotton Turtleneck Sweater/i })
   ).toBeVisible();
 });
+
+// TODO: add test for adding product to wishlist
